Move tests from legs_test.go into other files
Use separate test package to test dagsync code.
